created basic layout of website
removed the dependency of mongo added mysql and basic look of the website is prepared
