+++ controllers/homeController.js
@@ -1,0 +1,5 @@
+const homeController = (req, res) => {
+    res.render('index')
+}
+
+export { homeController }
