+++ package.json
@@ -9,10 +9,10 @@
   "author": "Vipin",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.20.2",
     "bootstrap": "^5.3.1",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
+    "express-ejs-layouts": "^2.5.1",
     "mysql": "^2.18.1",
     "request": "^2.88.2"
   }
