+++ public/css/signup.css
@@ -1,0 +1,47 @@
+
+.form-signin {
+  width: 100%;
+  max-width: 330px;
+  padding: 15px;
+  margin: 0 auto;
+}
+.form-signin .checkbox {
+  font-weight: 400;
+}
+.form-signin .form-control {
+  position: relative;
+  box-sizing: border-box;
+  height: auto;
+  padding: 10px;
+  font-size: 16px;
+}
+.form-signin .form-control:focus {
+  z-index: 2;
+}
+.form-signin input[id="fname"] {
+  margin-bottom: -1px;
+  border-bottom-left-radius: 0;
+  border-bottom-right-radius: 0;
+}
+.form-signin input[id="lname"] {
+  margin-bottom: -1px;
+  border-radius: 0;
+}
+.form-signin input[type="email"] {
+  margin-bottom: -1px;
+  border-radius: 0;
+}
+.form-signin input[id="inputPassword"] {
+  border-radius: 0;
+  margin-bottom: -1px;
+}
+.form-signin input[id="confirmPassword"] {
+  border-top-left-radius: 0;
+  border-top-right-radius: 0;
+  margin-bottom: 10px;
+}
+#submitbtn{
+  width: 100%;
+  display: block;
+  margin: auto;
+}
