--- public/css/styles.css
@@ -1,67 +1,0 @@
-html,
-body {
-  height: 100%;
-  background-color: #96C291 !important;
-}
-
-body {
-  display: -ms-flexbox;
-  display: -webkit-box;
-  display: flex;
-  -ms-flex-align: center;
-  -ms-flex-pack: center;
-  -webkit-box-align: center;
-  align-items: center;
-  -webkit-box-pack: center;
-  justify-content: center;
-  padding-top: 40px;
-  padding-bottom: 40px;
-  background-color: #f5f5f5;
-}
-
-.form-signin {
-  width: 100%;
-  max-width: 330px;
-  padding: 15px;
-  margin: 0 auto;
-}
-.form-signin .checkbox {
-  font-weight: 400;
-}
-.form-signin .form-control {
-  position: relative;
-  box-sizing: border-box;
-  height: auto;
-  padding: 10px;
-  font-size: 16px;
-}
-.form-signin .form-control:focus {
-  z-index: 2;
-}
-.form-signin input[id="fname"] {
-  margin-bottom: -1px;
-  border-bottom-left-radius: 0;
-  border-bottom-right-radius: 0;
-}
-.form-signin input[id="lname"] {
-  margin-bottom: -1px;
-  border-radius: 0;
-}
-.form-signin input[type="email"] {
-  margin-bottom: -1px;
-  border-radius: 0;
-}
-.form-signin input[id="inputPassword"] {
-  border-radius: 0;
-  margin-bottom: -1px;
-}
-.form-signin input[id="confirmPassword"] {
-  border-top-left-radius: 0;
-  border-top-right-radius: 0;
-  margin-bottom: 10px;
-}
-#submitbtn{
-  width: 100%;
-  display: block;
-  margin: auto;
-}
