--- views/index.ejs
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Welcome to My Website</title>
-</head>
-<body>
-    <h1>Welcome to My Website</h1>
-    <p>This is the homepage content.</p>
-</body>
-</html>
