+++ views/login.ejs
@@ -1,0 +1,20 @@
+<%- include('partials/nav_other') %>
+<div id="login" class="p-5" >
+  <div id="innerDiv" style="border: solid 1px gray ; background-color: rgb(190, 190, 190); width: 30vw;" class="m-auto rounded p-5 text-center">
+    <main>
+
+      <form class="form-login w-100">
+        <img class="mb-4" src="images/logo.png" alt="" width="72" height="72">
+        <h1 class="h3 mb-3 font-weight-normal">Login</h1>
+        <!-- <label for="inputEmail" class="sr-only">Email address</label> -->
+        <input type="email" id="inputEmail" class="form-control" placeholder="Email address" required autofocus>
+        <!-- <label for="inputPassword" class="sr-only">Password</label> -->
+        <input type="password" id="inputPassword" class="form-control" placeholder="Password" required>
+        <button id="sub_btn" class="btn btn-primary btn-lg btn-block" type="submit">Login</button>
+      </form>
+      <p class="text-center">Don't have an account, <a class="link" href="/signup">Signup</a></p>
+    </div>  
+  </main>
+</div>
+</div>
+<%- include('partials/footer') %>
