+++ views/mygallery.ejs
@@ -1,0 +1,144 @@
+<%- include('partials/nav_other') %>
+  
+  <!-- Search Bar -->
+  <div class="row col-8 m-auto">
+  <div class="input-group mt-3 col-6">
+  <input type="search" class="form-control rounded-left" placeholder="Search" aria-label="Search" aria-describedby="search-addon" />
+  <button type="button" class="btn btn-success">search</button>
+  </div>
+  </div>
+  
+  <!-- Gallery -->
+  <div class="row col-8 mt-3 m-auto">
+  <div class="col-lg-4 col-md-12 mb-4 mb-lg-0">
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(73).webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Boat on Calm Water"
+  />
+  
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain1.webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Wintry Mountain Landscape"
+  />
+  </div>
+  
+  <div class="col-lg-4 mb-4 mb-lg-0">
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain2.webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Mountains in the Clouds"
+  />
+  
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(73).webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Boat on Calm Water"
+  />
+  </div>
+  
+  <div class="col-lg-4 mb-4 mb-lg-0">
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(18).webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Waves at Sea"
+  />
+  
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain3.webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Yosemite National Park"
+  />
+  </div>
+  </div>
+  <!-- Gallery -->
+  <div class="row col-8 mt-3 m-auto">
+  <div class="col-lg-4 col-md-12 mb-4 mb-lg-0">
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(73).webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Boat on Calm Water"
+  />
+  
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain1.webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Wintry Mountain Landscape"
+  />
+  </div>
+  
+  <div class="col-lg-4 mb-4 mb-lg-0">
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain2.webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Mountains in the Clouds"
+  />
+  
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(73).webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Boat on Calm Water"
+  />
+  </div>
+  
+  <div class="col-lg-4 mb-4 mb-lg-0">
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(18).webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Waves at Sea"
+  />
+  
+  <img
+    src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain3.webp"
+    class="w-100 shadow-1-strong rounded mb-4"
+    alt="Yosemite National Park"
+  />
+  </div>
+  </div>
+  <!-- Gallery -->
+  <div class="row col-8 mt-3 m-auto">
+    <div class="col-lg-4 col-md-12 mb-4 mb-lg-0">
+    <img
+      src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(73).webp"
+      class="w-100 shadow-1-strong rounded mb-4"
+      alt="Boat on Calm Water"
+    />
+  
+    <img
+      src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain1.webp"
+      class="w-100 shadow-1-strong rounded mb-4"
+      alt="Wintry Mountain Landscape"
+    />
+  </div>
+  
+  <div class="col-lg-4 mb-4 mb-lg-0">
+    <img
+      src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain2.webp"
+      class="w-100 shadow-1-strong rounded mb-4"
+      alt="Mountains in the Clouds"
+    />
+  
+    <img
+      src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(73).webp"
+      class="w-100 shadow-1-strong rounded mb-4"
+      alt="Boat on Calm Water"
+    />
+  </div>
+  
+  <div class="col-lg-4 mb-4 mb-lg-0">
+    <img
+      src="https://mdbcdn.b-cdn.net/img/Photos/Horizontal/Nature/4-col/img%20(18).webp"
+      class="w-100 shadow-1-strong rounded mb-4"
+      alt="Waves at Sea"
+    />
+  
+    <img
+      src="https://mdbcdn.b-cdn.net/img/Photos/Vertical/mountain3.webp"
+      class="w-100 shadow-1-strong rounded mb-4"
+      alt="Yosemite National Park"
+    />
+  </div>
+  </div>
+
+<%- include('partials/footer') %>
