+++ views/partials/nav_home.ejs
@@ -1,0 +1,27 @@
+<nav style="position: absolute;" id="nav_other" class="navbar navbar-expand-lg fixed-top bg-transparent">
+    <div class="container">
+      <a class="navbar-brand text-white fw-bold" id="logo" href="#">PicGallery</a>
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarNav">
+      <ul class="nav ms-auto ">
+        <li class="nav-item active">
+          <a class="nav-link text-white" href="/home">Home</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/about">About</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/upload">Upload</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/mygallery">Gallery</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/login">Login</a>
+        </li>
+      </ul>
+    </div>
+  </div>
+  </nav>
