+++ views/partials/nav_other.ejs
@@ -1,0 +1,28 @@
+<nav id="nav_other" class="navbar navbar-expand-lg navbar-dark bg-dark">
+  
+    <div class="container">
+      <a class="navbar-brand text-white fw-bold" id="logo" href="#">PicGallery</a>
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarNav">
+      <ul class="nav ms-auto ">
+        <li class="nav-item active">
+          <a class="nav-link text-white" href="/home">Home</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/about">about</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/upload">upload</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/mygallery">Gallery</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-white" href="/login">Login</a>
+        </li>
+      </ul>
+    </div>
+  </div>
+  </nav>
